Adding emac_pins_b to device tree for sun7i-a20
The A20 has a pin muxing option that allows the EMAC pins to be assigned to either Port A or Port H.
Adding the list for the Port H variant. This may or may not work with the existing driver.
On Port A, the pins are Multiplex Group 2 for EMAC, Multiplex group 5 for GMAC.
On Port A, the pins are Multiplex Group 3 for EMAC.
Note that using the EMAC on port H removes access to LCD1, Keypad, CAN, SD/MMC #1, and CSI1.